# src/config/swagger/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Perfume Shop API",
    "description": "RESTful API for a perfume and cologne e-commerce platform with multi-variant product support, allowing flexible volume-based pricing and weight-based sales",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Development server"
    },
    {
      "url": "https://api.yourdomain.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/{slug}/reactivate": {
      "patch": {
        "tags": [
          "Admin - Categories"
        ],
        "summary": "Reactivate a category",
        "description": "Activate a deactivated category",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Category slug"
          }
        ],
        "responses": {
          "200": {
            "description": "Category reactivated successfully"
          },
          "400": {
            "description": "Category already active"
          },
          "401": {
            "description": "Authentication required or not admin"
          },
          "404": {
            "description": "Category not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/{userId}/addresses": {
      "get": {
        "tags": [
          "Admin - Users"
        ],
        "summary": "Get user addresses",
        "description": "Retrieve all addresses for a specific user (Admin only)",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID (MongoDB ObjectId)",
            "example": "507f1f77bcf86cd799439011"
          }
        ],
        "responses": {
          "200": {
            "description": "User addresses retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "username": {
                          "type": "string"
                        },
                        "addresses": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "_id": {
                                "type": "string"
                              },
                              "phone": {
                                "type": "string"
                              },
                              "addressLine": {
                                "type": "string"
                              },
                              "postalCode": {
                                "type": "string"
                              },
                              "stateId": {
                                "type": "integer"
                              },
                              "cityId": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or not admin"
          },
          "404": {
            "description": "User not found"
          },
          "409": {
            "description": "Invalid user ID format"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/{userId}/role": {
      "patch": {
        "tags": [
          "Admin - Users"
        ],
        "summary": "Change user role",
        "description": "Change user role to ADMIN or USER (Admin only)",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User role updated successfully"
          },
          "401": {
            "description": "Authentication required or not admin"
          },
          "404": {
            "description": "User not found"
          },
          "409": {
            "description": "Invalid user ID format"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "role": {
                    "type": "string",
                    "enum": [
                      "ADMIN",
                      "USER"
                    ],
                    "example": "ADMIN",
                    "description": "New user role"
                  }
                },
                "required": [
                  "role"
                ]
              }
            }
          }
        }
      }
    },
    "/{userId}/ban": {
      "patch": {
        "tags": [
          "Admin - Users"
        ],
        "summary": "Ban a user",
        "description": "Ban a user account (Admin only)",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User banned successfully"
          },
          "401": {
            "description": "Authentication required or not admin"
          },
          "404": {
            "description": "User not found"
          },
          "409": {
            "description": "Invalid user ID format"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/{userId}/unban": {
      "patch": {
        "tags": [
          "Admin - Users"
        ],
        "summary": "Unban a user",
        "description": "Remove ban from a user account (Admin only)",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User unbanned successfully"
          },
          "401": {
            "description": "Authentication required or not admin"
          },
          "404": {
            "description": "User not found"
          },
          "409": {
            "description": "Invalid user ID format"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/{userId}/reactivate": {
      "patch": {
        "tags": [
          "Admin - Users"
        ],
        "summary": "Reactivate a user",
        "description": "Reactivate a deactivated user account (Admin only)",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User reactivated successfully"
          },
          "401": {
            "description": "Authentication required or not admin"
          },
          "404": {
            "description": "User not found"
          },
          "409": {
            "description": "Invalid user ID format"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token"
      }
    }
  }
}